+++ src/module.c
@@ -11,6 +11,7 @@
 #include "playlistcontainer.h"
 #include "search.h"
 #include "session.h"
+#include "toplistbrowser.h"
 #include "track.h"
 #include "image.h"
 #include "user.h"
@@ -43,6 +44,8 @@ init_spotify(void)
         return;
     if (PyType_Ready(&ResultsType) < 0)
         return;
+    if (PyType_Ready(&ToplistBrowserType) < 0)
+        return;
     if (PyType_Ready(&TrackType) < 0)
         return;
     if (PyType_Ready(&ImageType) < 0)
@@ -73,6 +76,7 @@ init_spotify(void)
     playlistcontainer_init(m);
     session_init(m);
     search_init(m);
+    toplistbrowser_init(m);
     track_init(m);
     image_init(m);
     user_init(m);