+++ docs/changes.rst
@@ -8,9 +8,11 @@ v1.4 (in development)
 New features:
 
 - User handling
+- Toplist browsing
 - New Playlist method: rename.
 - Exceptions raised in callbacks are written to ``stderr``.
 - Unicode queries in Session.search
+- New Session method: get_friends.
 
 Bug fixes: